requirements.txt
tweepy
pillow
requests
schedule